requirements.txt
pandas
spacy
tqdm
nltk
pdfminer.six
openpyxl
python-docx
